authorization complete

+++ src/api/notes/handler.js
@@ -16,9 +16,12 @@ class NotesHandler {
     async postNoteHandler(request, h) {
         try {
             this._validator.validateNotePayload(request.payload);
-            const {title = 'untitled', body, tags} = request.payload;
+            const { title = 'untitled', body, tags } = request.payload;
+            const { id: credentialId } = request.auth.credentials;
 
-            const noteId = await this._service.addNote({title, body, tags});
+            const noteId = await this._service.addNote({
+                title, body, tags, owner: credentialId,
+            });
 
             const response = h.response({
                 status: 'success',
@@ -49,9 +52,10 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler(h) {
+    async getNotesHandler(request, h) {
         try {
-            const notes = await this._service.getNotes();
+            const { id :credentialId } = request.auth.credentials;
+            const notes = await this._service.getNotes(credentialId);
             return {
                 status: 'success',
                 data: {
@@ -81,7 +85,11 @@ class NotesHandler {
 
     async getNoteByIdHandler(request, h) {
         try {
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId)
+
             const note = await this._service.getNoteById(id);
             return {
                 status: 'success',
@@ -113,10 +121,11 @@ class NotesHandler {
     async putNoteByIdHandler(request, h) {
         try {
             this._validator.validateNotePayload(request.payload);
-            const {title, body, tags}  = request.payload;
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
 
-            await this._service.editNoteById(id, {title, body, tags});
+            await this._service.verifyNoteOwner(id, credentialId);
+            await this._service.editNoteById(id, request.payload);
 
             return {
                 status: 'success',
@@ -146,8 +155,10 @@ class NotesHandler {
 
     async deleteNoteByIdHandler(request, h) {
         try {
-            const {id} = request.params;
+            const { id } = request.params;
+            const { id: credentialId } = request.auth.credentials;
 
+            await this._service.verifyNoteOwner(id, credentialId);
             await this._service.deleteNoteById(id);
 
             return {
